Add files via upload
The Walmart Superstore wants to analyze its sales data to gain insights into its performance and identify opportunities for growth. The company has a vast amount of sales data, including information on products, customers, stores, and transactions. The main objective is to extract meaningful insights from this data to analyse the overall sales of walmart superstore.